+++ components/services/MvpDevelopment.tsx
@@ -1,8 +1,7 @@
 import { Metadata } from "next";
-import { useTranslations } from "next-intl";
 import { getTranslations } from "next-intl/server";
-import { RocketLaunchIcon, CheckIcon } from "@heroicons/react/24/outline";
-import { Button } from "@/components/ui/button";
+import { useTranslations } from "next-intl";
+import { CheckIcon, ChartPieIcon } from "@heroicons/react/24/outline";
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
 import { Badge } from "@/components/ui/badge";
 import { getFeatureGroups, getBenefits, technologies, getWorkTypes } from "@/content/data";
@@ -11,16 +10,16 @@ import { Link } from "@/i18n/navigation";
 export async function generateMetadata({ params }: { params: Promise<{locale: string}> }): Promise<Metadata> {
   const {locale} = await params;
   const t = await getTranslations({ locale, namespace: "metadata" });
-  const rawKeywords = t.raw("businessWeb.keywords");
+  const rawKeywords = t.raw("mvp.keywords");
 
   return {
-    title: t("businessWeb.title"),
-    description: t("businessWeb.description"),
+    title: t("mvp.title"),
+    description: t("mvp.description"),
     keywords: rawKeywords
   };
 }
 
-export default function BusinessWebsitesPlatforms() {
+export default function MvpDevelopmentPage() {
     const t = useTranslations();
     const features = getFeatureGroups(t);
     const benefits = getBenefits(t);
@@ -32,23 +31,23 @@ export default function BusinessWebsitesPlatforms() {
                 <div className="container mx-auto">
                     <div className="max-w-4xl mx-auto text-center">
                         <div className="size-20 bg-background rounded-2xl flex items-center justify-center mx-auto mb-8 border border-border">
-                            <RocketLaunchIcon className="size-10 text-orange-400" />
+                            <ChartPieIcon className="size-10 text-green-400" />
                         </div>
                         <h1 className="text-4xl sm:text-5xl font-bold text-foreground mb-6 font-poppins">
-                        { t("Service_BusinessWeb.hero.heading") } <span className="bg-gradient-to-r from-orange-400 to-red-400 bg-clip-text text-transparent">{ t("Service_BusinessWeb.hero.heading_highlighted") }</span>
+                        { t("Service_MVP.hero.heading") } <span className="bg-gradient-to-r from-green-400 to-lime-400 bg-clip-text text-transparent">{ t("Service_MVP.hero.heading_highlighted") }</span>
                         </h1>
                         <p className="text-lg text-muted-foreground mb-8 max-w-2xl mx-auto">
-                            { t("Service_BusinessWeb.hero.subheading") }
+                            { t("Service_MVP.hero.subheading") }
                         </p>
                         <div className="flex flex-wrap gap-2 justify-center mb-8">
-                            {technologies.business_websites_platforms.map((tech) => (
+                            {technologies.mvp_development.map((tech) => (
                                 <Badge key={tech} variant="secondary" className="bg-card border border-border">
                                     {tech}
                                 </Badge>
                             ))}
                         </div>
-                        <Link href={{ pathname: "/", hash: "contact" }} className="bg-orange-400 hover:bg-orange-400/90 hover:cursor-pointer py-3 font-semibold px-5 rounded-lg transition-colors duration-300 ease-in-out block mt-5 md:w-fit mx-auto text-sm">
-                            { t("Service_BusinessWeb.hero.cta") }
+                        <Link href={{ pathname: "/", hash: "contact" }} className="bg-green-600 hover:bg-green-600/90 hover:cursor-pointer py-3 font-semibold px-5 rounded-lg transition-colors duration-300 ease-in-out block mt-5 md:w-fit mx-auto text-sm">
+                            { t("Service_MVP.hero.cta") }
                         </Link>
                     </div>
                 </div>
@@ -58,19 +57,19 @@ export default function BusinessWebsitesPlatforms() {
                 <div className="container mx-auto">
                     <div className="text-center mb-16">
                         <h2 className="text-3xl font-bold text-foreground mb-4 font-poppins">
-                            { t("Service_BusinessWeb.features_section.heading") } <span className="bg-gradient-to-r from-orange-400 to-red-400 bg-clip-text text-transparent">{ t("Service_BusinessWeb.features_section.heading_highlighted") }</span>
+                            { t("Service_MVP.features_section.heading") } <span className="bg-gradient-to-r from-green-400 to-lime-400 bg-clip-text text-transparent">{ t("Service_MVP.features_section.heading_highlighted") }</span>
                         </h2>
                         <p className="text-lg text-muted-foreground max-w-2xl mx-auto">
-                            { t("Service_BusinessWeb.features_section.subheading") }
+                            { t("Service_MVP.features_section.subheading") }
                         </p>
                     </div>
 
                     <div className="grid grid-cols-1 md:grid-cols-2 gap-8 mb-16">
-                        {features.business_websites_platforms.map((feature, index) => (
+                        {features.mvp_development.map((feature, index) => (
                             <Card key={index} className="bg-card border-border">
                                 <CardHeader>
-                                    <div className="size-12 bg-orange-400/10 rounded-lg flex items-center justify-center mb-4">
-                                        <feature.icon className="size-7 text-orange-400" />
+                                    <div className="size-12 bg-green-400/10 rounded-lg flex items-center justify-center mb-4">
+                                        <feature.icon className="size-7 text-green-400" />
                                     </div>
                                 <CardTitle className="text-xl text-card-foreground font-poppins font-medium">{feature.title}</CardTitle>
                                 <CardDescription className="text-muted-foreground">{feature.description}</CardDescription>
@@ -83,25 +82,16 @@ export default function BusinessWebsitesPlatforms() {
 
             <section className="py-16 px-4 sm:px-6 lg:px-8">
                 <div className="container mx-auto">
-                    <h2 className="text-3xl font-bold text-center text-foreground mb-12 font-poppins">{ t("Service_BusinessWeb.types.heading") }</h2>
-                    <div className="grid grid-cols-1 md:grid-cols-3 gap-8">
-                        {workTypes.business_websites_platforms.map((type, index) => (
-                            <Card key={index} className="bg-card border-border">
-                                <CardHeader>
-                                    <CardTitle className="text-xl text-card-foreground font-poppins font-medium">{type.title}</CardTitle>
-                                    <CardDescription>{type.description}</CardDescription>
-                                </CardHeader>
-                                <CardContent>
-                                <ul className="space-y-2">
-                                    {type.features.map((feature, featureIndex) => (
-                                    <li key={featureIndex} className="flex items-center text-sm text-muted-foreground">
-                                        <div className="size-1.5 bg-orange-400 rounded-full mr-3" />
-                                        {feature}
-                                    </li>
-                                    ))}
-                                </ul>
-                                </CardContent>
-                            </Card>
+                    <h2 className="text-3xl text-center text-foreground mb-12 font-poppins font-bold">{ t("Service_MVP.process.heading") }</h2>
+                    <div className="grid grid-cols-1 md:grid-cols-4 gap-8">
+                        {workTypes.mvp_development.map((phase, index) => (
+                            <div key={index} className="text-center">
+                                <div className="w-16 h-16 bg-green-600 text-white rounded-full flex items-center justify-center mx-auto mb-4 text-xl font-bold">
+                                    {phase.step}
+                                </div>
+                                <h3 className="text-xl font-semibold text-foreground mb-2 font-poppins">{phase.title}</h3>
+                                <p className="text-muted-foreground">{phase.description}</p>
+                            </div>
                         ))}
                     </div>
                 </div>
@@ -111,11 +101,11 @@ export default function BusinessWebsitesPlatforms() {
                 <div className="container mx-auto">
                     <div className="grid grid-cols-1 lg:grid-cols-2 gap-12 items-center">
                         <div>
-                            <h3 className="text-2xl font-bold text-foreground mb-6 font-poppins">{ t("Service_BusinessWeb.benefits.heading") }</h3>
+                            <h3 className="text-2xl font-bold text-foreground mb-6 font-poppins">{ t("Service_MVP.benefits.heading") }</h3>
                             <div className="space-y-4">
                                 {benefits.business_websites_platforms.map((benefit, index) => (
                                 <div key={index} className="flex items-center">
-                                    <CheckIcon className="size-5 text-orange-400 mr-3 flex-shrink-0" />
+                                    <CheckIcon className="size-5 text-green-400 mr-3 flex-shrink-0" />
                                     <span className="text-muted-foreground">{benefit}</span>
                                 </div>
                                 ))}
@@ -123,13 +113,13 @@ export default function BusinessWebsitesPlatforms() {
                         </div>
                         <Card className="bg-card border-border">
                             <CardHeader>
-                                <CardTitle className="text-card-foreground font-poppins">{ t("Service_BusinessWeb.benefits.card.title") }</CardTitle>
+                                <CardTitle className="text-card-foreground font-poppins">{ t("Service_MVP.benefits.card.title") }</CardTitle>
                                 <CardDescription className="text-muted-foreground">
-                                    { t("Service_BusinessWeb.benefits.card.description") }
+                                    { t("Service_MVP.benefits.card.description") }
                                 </CardDescription>
                             </CardHeader>
                             <CardContent>
-                                <Link href={{ pathname: "/", hash: "contact" }} className="block w-full bg-orange-400 hover:bg-orange-400/90 hover:cursor-pointer text-md text-center py-2 rounded-md font-semibold transition-colors duration-300 ease-in-out">{ t("Service_BusinessWeb.benefits.card.cta") }</Link>
+                                <Link href={{ pathname: "/", hash: "contact" }} className="block w-full bg-green-600 hover:bg-green-600/90 hover:cursor-pointer text-md text-center py-2 rounded-md font-semibold transition-colors duration-300 ease-in-out">{ t("Service_BusinessWeb.benefits.card.cta") }</Link>
                             </CardContent>
                         </Card>
                     </div>
